requirements.txt
requests>=2.31.0
mcstatus>=11.0.0
tkinter-async>=0.2.0
Pillow>=10.0.0
geoip2>=4.7.0
maxminddb>=2.2.0
python-nmap>=0.7.1
threading-helper>=0.1.0
ipaddress>=1.0.23
